Add a parser to separate queries
Currently, we consider the whole migration file as a single query. It doesn't
help identify specific errors when one particular query is failing in a long
migration script. To be able to single out troublesome queries, we need to run
them sequentially and separately. For this, we'll need a simple parser that
will separate queries. This could probably be done through attoparsec or some
parsing library, but the logic should be simple enough so we can write it
manually.